点赞模块 - 使用 Redis 替代数据库 + 定时任务同步到数据库

+++ src/main/java/com/easen/aicode/model/vo/AppThumbVO.java
@@ -28,5 +28,5 @@ public class AppThumbVO implements Serializable {
     /**
      * 点赞数
      */
-    private Long thumbCount;
+    private Integer thumbCount;
 }
